Add `username` and `password` to Calix class initialization
The `username` and `password` parameters have been added to the init function of the Calix class, allowing for token retrieval upon class instantiation. This eliminates the need to explicitly call the token method from outside of the class. In addition, a method called `devices` has been added for retrieving and handling device data from the Calix API.